fix: Keep track of customized callouts w/o source
When a user configures a callout (e.g. from a snippet) and then removes
the source of the callout, Callout Manager will still apply styles for
it without keeping it inside the callout collection. This can be
confusing for users.

This commit makes it so Callout Manager will add undetected-but-modified
callouts as custom callouts, which the user can then choose to remove
manually through the UI.